# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: E-commerce API
  description: E-commerce API with support for online payments, product and order administration
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1'
    description: Development server
...
paths:
...
  /products:
    description: This resource represents products of the system.
    get:
      tags:
        - Products
      description:
        View all products
      responses:
        '200':
          description: View of all products
          content:
            application/json:
              schema:
                $ref: ./product.json#/definitions/productsAll
              example:
                - ID: 1
                  title: Product 1
                  description: description
                  image: https://pixabay.com/photos/tree-sunset-clouds-sky-silhouette-736885/
                  price: 12.99
                  links: 
                    self: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products"
                    detail: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/1"
                    update: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/1"
                    delete: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/1"
                    add: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products"
                - ID: 2
                  title: Product 2
                  description: description
                  image: https://pixabay.com/photos/tree-sunset-clouds-sky-silhouette-736885/
                  price: 10.99
                  links: 
                    self: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products"
                    detail: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/2"
                    update: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/2"
                    delete: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/2"
                    add: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products"
    post:
      tags:
        - Products
      description:
        Create a new product
      requestBody:
        description: Product data
        required: true
        content:
          application/json:
            schema:
              $ref: ./product.json#/definitions/productCreate 
            example:
              title: Product 1
              description: description
              image: https://pixabay.com/photos/tree-sunset-clouds-sky-silhouette-736885/
              price: 12.99
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: ./product.json#/definitions/product
              example:
                ID: 1
                title: Product 1
                description: description
                image: https://pixabay.com/photos/tree-sunset-clouds-sky-silhouette-736885/
                price: 12.99
                links: 
                    self: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products"
                    update: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/1"
                    delete: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products/1"
                    allProducts: "http://goodvertigo-chariotclarion-3000.codio-box.uk/api/v1/products"
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: requires property "price"
        '401':
          description: Only admin users can add new products
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Permission not granted
  /products/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Product id
    get:
      tags:
        - Products
      description:
        View a product by id
      responses:
        '200':
          description: View of a product
          content:
            application/json:
              schema:
                $ref: ./product.json#/definitions/product
              example:
                ID: 1
                title: Product 1
                description: description
                image: https://pixabay.com/photos/tree-sunset-clouds-sky-silhouette-736885/
                price: 12.99
        '404':
          description: Product does not exist
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Product does not exist
    put:
      tags:
        - Products
      description:
        Update a product by id
      requestBody:
        description: Product data
        required: true
        content:
          application/json:
            schema:
              $ref: ./product.json#/definitions/productUpdate 
            example:
              title: Product 1
              description: description
              image: https://pixabay.com/photos/tree-sunset-clouds-sky-silhouette-736885/
              price: 12.99
      responses:
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: 1 product updated
        '400':
          description: Product could not be updated
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Product was not updated
        '401':
          description: Only admin users can update products
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Permission not granted
    delete:
      tags:
        - Products
      description:
        Delete a product by id
      responses:
        '200':
          description: Deletion successful
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: 1 product deleted
        '400':
          description: Product could not be deleted
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Product was not deleted
        '401':
          description: Only admin users can delete products
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Permission not granted
...
  /categories/{cid}/product/{pid}:
    parameters:
      - in: path
        name: cid
        schema:
          type: integer
        required: true
        description: Category id
      - in: path
        name: pid
        schema:
          type: integer
        required: true
        description: Product id
    post:
      tags:
        - Categories
      description:
        Add a product to a category
      responses:
        '201':
          description: Product added to the category
          content:
            application/json:
              schema:
                $ref: ./category.json#/definitions/categoryItem
              example:
                ID: 3
                productId: 2
                categoryId: 1
        '400':
          description: Product could not be added
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Cannot add or update a child row...
        '401':
          description: Only admin users can add products to categories
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Permission not granted
    delete:
      tags:
        - Categories
      description:
        Delete a product from a category
      responses:
        '200':
          description: Deletion successful
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: 1 product deleted from category 2
        '400':
          description: Product could not be deleted from the category
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Product was not deleted
        '401':
          description: Only admin users can delete categories
          content:
            application/json:
              schema:
                $ref: ./message.json#/definitions/message
              example:
                message: Permission not granted
...